Changed should only convert new art, art 20201112

+++ .github/actions/convertArt/entrypoint.sh
@@ -1,34 +1,40 @@
 #!/bin/sh
 
 # go into uimodule folder
-cd uimodule/webapp/art
+cd uimodule/webapp/art || exit
 
 # remove old temporary folder (if available)
-rm -Rf .temp 2>&1 >/dev/null
+rm -Rf .temp >/dev/null 2>&1
 
-echo -n "[" > list.json
+echo -n "[" >list.json
 
 cnt=0
 for i in *.kra; do
-    cnt=$((cnt+1))
+    cnt=$((cnt + 1))
     echo "${cnt}: processing ${i}"
-    mkdir .temp
-    cd .temp
-    unzip -qq ../"${i}"
-    if [ -e "mergedimage.png" ]; then
-        convert -trim "mergedimage.png" -border 10x10 "../${i}.png"
-        if [ -e "preview.png" ]; then
-            convert -trim "preview.png" -border 10x10 "../${i}_thumb.png"
+    exportName="../${i}.png"
+    exportNameThumb="../${i}_thumb.png"
+    if [ ! -e "$exportName" ]; then
+        mkdir .temp
+        cd .temp || exit
+        unzip -qq ../"${i}"
+        if [ -e "mergedimage.png" ]; then
+            convert -trim "mergedimage.png" -border 10x10 "$exportName"
+            if [ -e "preview.png" ]; then
+                convert -trim "preview.png" -border 10x10 "$exportNameThumb"
+            fi
+        else
+            echo "${cnt}: ${i}: cannot find mergedimage.png!"
         fi
+        if [ "${cnt}" -gt "1" ]; then
+            echo -n "," >>../list.json
+        fi
+        echo "\"../${i}.png\"" >>../list.json
+        cd ..
+        rm -Rf .temp
     else
-        echo "${cnt}: ${i}: cannot find mergedimage.png!"
+        echo "skipped because it already exists."
     fi
-    if [ "${cnt}" -gt "1" ]; then
-        echo -n "," >> ../list.json
-    fi
-    echo "\"../${i}.png\"" >> ../list.json
-    cd ..
-    rm -Rf .temp
 done
 
-echo -n "]" >> list.json
+echo -n "]" >>list.json
